requirements.txt
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7
pyngrok==7.0.0
google-generativeai==0.3.1
transformers==4.35.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
faiss-cpu==1.7.4
scikit-learn==1.3.0
pandas==2.0.3
matplotlib==3.7.2
tqdm==4.66.1
underthesea==6.5.0
pyvi==0.1.1
rank_bm25==0.2.2
torch==2.0.1
supabase-py==2.3.0
psycopg2-binary==2.9.9
supabase